Update example/js files.
