+++ example/js/data/collection.json
